gaussian_near_field_ml

Electric field of a gaussian beam close to a planar interface

gaussian_near_field_ml(xyz, wavelength = 632.8, alpha = 15 * pi/180,
  psi = 0, w0 = 10000, epsilon = c(1.5^2, epsAg(wavelength)$epsilon, 1^2,
  1^2), thickness = c(0, 50, 10, 0), maxEval = 3000, reqAbsError = 0,
  tol = 1e-04, progress = FALSE, field = FALSE)

Arguments

xyz

position matrix

wavelength

wavelength

alpha

beam incident angle

psi

beam polarisation angle

w0

beam waist radius

epsilon

vector of permittivities

thickness

thickness corresponding to each medium

maxEval

passed to cubature

reqAbsError

passed to cubature

tol

passed to cubature

progress

logical display progress bar

field

logical: return the electric field (complex vector), or modulus squared

Value

data.frame electric field at the x, y, z position

Details

Integration is performed over a spectrum of incident plane waves using integrand_gb2

See also

Other gaussian_beam: gaussian_near_field_layer